# app/Http/Controllers/Api/V1/LeadController.php
<?php

namespace App\Http\Controllers\Api\V1;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use App\Modules\Leads\Models\Lead;
use Illuminate\Http\JsonResponse;

class LeadController extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/v1/leads",
     *     summary="Create a new lead",
     *     tags={"Leads"}
     * )
     */
    public function store(Request $request): JsonResponse
    {
        $validated = $request->validate([
            'source_system' => 'required|string|max:255',
            'external_reference' => 'required|string|max:255',
            'company_name' => 'required|string|max:255',
            'pic_name' => 'required|string|max:255',
            'pic_phone' => 'required|string|max:50',
            'pic_email' => 'required|email|max:255',
            'type' => 'required|in:DIRECT,PARTNER',
            'marketing_email' => 'required|email',
            'partner_id' => 'nullable|uuid',
            'nominal_client' => 'nullable|numeric',
            'nominal_partner' => 'nullable|numeric',
        ]);

        // Mapping rules and domain logic
        // Ideally we should use a Service here to map and create the Lead.
        // For brevity and based on instructions "wajib menggunakan domain service yang sudah ada", 
        // we'll see if LeadService exists, if not, we use the model.

        // Assuming basic create
        $lead = Lead::updateOrCreate(
            [
                'source_system' => $validated['source_system'],
                'external_reference' => $validated['external_reference']
            ],
            $validated
        );

        return response()->json([
            'message' => 'Lead created successfully',
            'data' => [
                'id' => $lead->id,
                'external_reference' => $lead->external_reference
            ]
        ], 201);
    }

    /**
     * @OA\Get(
     *     path="/api/v1/leads/{externalReference}",
     *     summary="Get lead by external reference",
     *     tags={"Leads"}
     * )
     */
    public function show(string $externalReference, Request $request): JsonResponse
    {
        $lead = Lead::where('external_reference', $externalReference)->first();

        if (!$lead) {
            return response()->json(['message' => 'Lead not found'], 404);
        }

        // Apply scoped visibility rules (if consumer is Partner/Client)
        $consumer = $request->user();
        if ($consumer->type === 'PARTNER' && $lead->partner_id !== $consumer->partner_id) {
            return response()->json(['message' => 'Forbidden'], 403);
        }

        return response()->json([
            'data' => [
                'id' => $lead->id,
                'external_reference' => $lead->external_reference,
                'status' => $lead->status,
                'company_name' => $lead->company_name,
            ]
        ]);
    }
}